Update text editor capability
It's now only missing some characters like ! and ?. Text automatically goes to the next line and shift and enter keys work. The user's file is also automatically turned into a txt file at the end so the user just needs to copy and paste their work into a word file.